# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: p(DOOM) public API
  version: 0.1.0
  summary: Read-only access to the promoted p(DOOM) release, plus the Scenario Lab and public submissions.
  description: |
    The AI Existential and Civilizational Risk Observatory publishes its data through this
    API. Every response is served from the **current promoted release** in `data/releases`
    and from the sealed data snapshot that release was computed from. Handlers never compute
    or alter an estimate.

    Guarantees:

    * **Release-backed.** Every data response embeds a `meta` object naming the release,
      the snapshot, the source cutoff, the model versions and the release's known
      limitations. Estimate objects always carry `status`, `producer`, `horizon`,
      `outcome_set` and `conditioning`.
    * **No official probability.** In this release line the official object has status
      `insufficiently_calibrated` and publishes no number. External forecast aggregates,
      the research-mode model, the 0–100 indexes and the editorial level are separate
      objects and are never blended. Indexes are not probabilities.
    * **No automatic publication.** Nothing reachable over HTTP changes a release.
      Submissions are appended to the human review queue only. The Scenario Lab returns
      results labelled `user_scenario` that are computed per request and stored nowhere.
    * **No administrative endpoints.** Promotion, rollback, review and key management are
...
  contact:
...
servers:
...
paths:
...
  /v1/meter:
    get:
      tags: [meter]
      summary: Current release summary
      description: |
        Manifest fields, the official object (status `insufficiently_calibrated`, no
        probability), external forecast aggregates, research-mode estimates, the seven
        indexes, the editorial risk level, the uncertainty score, data cutoff, last review
        and publication timestamps, model versions and known limitations.
      operationId: getMeter
      responses:
        "200":
          description: The current release.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeterResponse"
        "304":
          description: Not modified.
        "503":
          $ref: "#/components/responses/Unavailable"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/meter/history:
    get:
      tags: [meter]
      summary: Release history
      description: |
        Every published release in id order with a per-release headline object. Directories
        under `data/releases` that cannot be read or are not marked published are omitted.
      operationId: getMeterHistory
      responses:
        "200":
          description: Headlines in chronological order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResponse"
        "503":
          $ref: "#/components/responses/Unavailable"
...
  /v1/drivers:
    get:
      tags: [data]
      summary: Driver families, signals, observations and the release's "why this number" rows
      operationId: getDrivers
      responses:
        "200":
          description: Drivers and observations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriversResponse"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/scenarios:
    get:
      tags: [data]
      summary: Scenarios S1–S18 and their edges
      operationId: listScenarios
      responses:
        "200":
          description: Scenarios.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScenariosResponse"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/scenarios/{id}:
    get:
      tags: [data]
      summary: One scenario with its edges and interventions
      operationId: getScenario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: "^S[0-9]{1,3}$"
          example: S3
      responses:
        "200":
          description: The scenario.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScenarioResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/forecasts:
    get:
      tags: [data]
      summary: Forecast records, compatibility-group aggregations and external aggregate estimates
      operationId: getForecasts
      responses:
        "200":
          description: Forecasts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForecastsResponse"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/capabilities:
    get:
      tags: [data]
      summary: Benchmarks and benchmark results
      operationId: getCapabilities
      responses:
        "200":
          description: Benchmarks.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CapabilitiesResponse"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/incidents:
    get:
      tags: [data]
      summary: Incident records (non-graphic, registry-referenced)
      operationId: getIncidents
      responses:
        "200":
          description: Incidents.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentsResponse"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/safeguards:
    get:
      tags: [data]
      summary: Interventions (safeguards)
      operationId: getSafeguards
      responses:
        "200":
          description: Interventions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafeguardsResponse"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/sources:
    get:
      tags: [data]
      summary: Source records with filters and pagination
      operationId: listSources
      parameters:
        - name: tier
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 5
          description: Exact source tier.
        - name: topic
          in: query
          schema:
            type: string
            maxLength: 100
          description: Case-insensitive exact match against the source's topic list.
        - name: q
          in: query
          schema:
            type: string
            maxLength: 200
          description: Case-insensitive substring over id, title, publisher, authors, citation, evidence summary and URL.
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: A page of sources.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourcesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/sources/{id}:
    get:
      tags: [data]
      summary: One source with its claims
      operationId: getSource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: "^src-[a-z0-9]+(-[a-z0-9]+)*$"
      responses:
        "200":
          description: The source.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/Unavailable"
...
  /v1/methodology/{slug}:
    get:
      tags: [reference]
      summary: One methodology document as Markdown
      operationId: getMethodologyDoc
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            pattern: "^[a-z0-9][a-z0-9_-]{0,63}$"
      responses:
        "200":
          description: The document.
          content:
            text/markdown:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/releases/{id}:
    get:
      tags: [releases]
      summary: One promoted release in full
      description: |
        Manifest, estimates, indexes, aggregations, sensitivity runs, delta record, drivers
        explained and approvals, exactly as promoted. Only published, readable release
        directories are served; any other id is `404`.
      operationId: getRelease
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: "^rel-[0-9]{4}-[0-9]{2}-[0-9]{2}-[0-9]{3}$"
      responses:
        "200":
          description: The release.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/releases/{id}/{doc}:
    get:
      tags: [releases]
      summary: Changelog or model card of a published release (Markdown)
      operationId: getReleaseDocument
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, pattern: "^rel-[0-9]{4}-[0-9]{2}-[0-9]{2}-[0-9]{3}$" }
        - name: doc
          in: path
          required: true
          schema: { type: string, enum: [changelog.md, model-card.md] }
      responses:
        "200":
          description: The document.
          content:
            text/markdown:
              schema: { type: string }
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/snapshot:
    get:
      tags: [data]
      summary: The sealed data snapshot behind the current release, entity file by entity file
      operationId: getSnapshot
      responses:
        "200":
          description: The snapshot. Large; cacheable by ETag.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotResponse"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/definitions:
    get:
      tags: [data]
      summary: Glossary definitions
      operationId: getDefinitions
      responses:
        "200":
          description: Definitions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefinitionsResponse"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/organizations:
    get:
      tags: [data]
      summary: Organizations
      operationId: getOrganizations
      responses:
        "200":
          description: Organizations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationsResponse"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/actions:
    get:
      tags: [data]
      summary: Audience-specific actions
      operationId: getActions
      responses:
        "200":
          description: Actions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionsResponse"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/scenario-lab/evaluate:
    post:
      tags: [scenario-lab]
      summary: Evaluate a user scenario
      description: |
        Runs the experimental research-mode model with the caller's sliders using the
        `model_spec.experimental_causal` specification of the **current** release's snapshot.
        The response is labelled `user_scenario`, carries the mandated disclaimer and is
        never an official estimate. The computation is deterministic for a given release,
        parameters, sample count and seed; nothing is stored. Rate limited (10 burst / 30 per minute).
      operationId: evaluateScenario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScenarioLabRequest"
      responses:
        "200":
          description: The user scenario result.
          headers:
            Cache-Control:
              schema:
                type: string
              description: Always `no-store`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScenarioLabResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/OriginNotAllowed"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/submissions/sources:
    post:
      tags: [submissions]
      summary: Suggest a source
      description: |
        Appends a `source` submission to `data/review-queue/submissions.jsonl` for human
        review. Never modifies snapshots or releases. Body ≤ 64 KiB; unknown fields are
        rejected; control characters are stripped. Rate limited (5 burst / 10 per minute).
      operationId: submitSource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceSubmissionRequest"
      responses:
        "202":
          description: Queued for review.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/OriginNotAllowed"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/submissions/corrections:
    post:
      tags: [submissions]
      summary: Suggest a correction
      description: |
        Appends a `correction` submission to `data/review-queue/submissions.jsonl` for human
        review. Never modifies snapshots or releases. Same limits as source submissions.
      operationId: submitCorrection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CorrectionSubmissionRequest"
      responses:
        "202":
          description: Queued for review.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/OriginNotAllowed"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "429":
          $ref: "#/components/responses/RateLimited"
components:
  headers:
    ETag:
      description: Strong validator, the SHA-256 of the body in quotes.
      schema:
        type: string
    CacheControl:
      description: "`public, max-age=60` on every GET."
      schema:
        type: string
  responses:
    BadRequest:
      description: The request is malformed; `detail` lists every problem found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No such resource.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    RateLimited:
      description: The per-client token bucket is empty; retry after `Retry-After` seconds.
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unavailable:
      description: No promoted release is loaded.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    MeterResponse:
      type: object
      required: [meta, manifest, official, external_aggregates, research_mode, indexes, editorial_risk_level, uncertainty_score, uncertainty_label, data_cutoff, last_reviewed, last_published, model_versions, limitations, headline]
      properties:
        meta: { $ref: "#/components/schemas/Meta" }
        manifest: { $ref: "#/components/schemas/ReleaseManifest" }
        official:
          type: array
          description: One object per horizon with status insufficiently_calibrated and null quantiles.
          items: { $ref: "#/components/schemas/Estimate" }
        external_aggregates:
          type: array
          items: { $ref: "#/components/schemas/Estimate" }
        research_mode:
          type: array
          items: { $ref: "#/components/schemas/Estimate" }
        indexes:
          type: array
          items: { $ref: "#/components/schemas/IndexValue" }
        editorial_risk_level: { type: string }
        uncertainty_score: { type: [number, "null"], description: "0–100 score; not a probability." }
        uncertainty_label: { $ref: "#/components/schemas/UncertaintyLabel" }
        data_cutoff: { type: string, format: date }
        last_reviewed: { type: [string, "null"], format: date-time, description: "Latest reviewer signature." }
        last_published: { type: [string, "null"], format: date-time }
        model_versions:
          type: array
          items: { type: string }
        limitations:
          type: array
          items: { type: string }
        headline: { $ref: "#/components/schemas/Headline" }
    HistoryResponse:
      type: object
      required: [meta, releases]
      properties:
        meta: { $ref: "#/components/schemas/Meta" }
        releases:
          type: array
          items: { $ref: "#/components/schemas/Headline" }
...
    DriversResponse:
      type: object
      required: [meta, drivers, observations, drivers_explained]
      properties:
        meta: { $ref: "#/components/schemas/Meta" }
        drivers:
          type: array
          items: { $ref: "#/components/schemas/Entity" }
        observations:
          type: array
          items: { $ref: "#/components/schemas/Entity" }
        drivers_explained:
          type: object
          additionalProperties: true
    ScenariosResponse:
      type: object
      required: [meta, scenarios, edges, note]
      properties:
        meta: { $ref: "#/components/schemas/Meta" }
        scenarios:
          type: array
          items: { $ref: "#/components/schemas/Entity" }
        edges:
          type: array
          items: { $ref: "#/components/schemas/Entity" }
        note: { type: string }
    ScenarioResponse:
      type: object
      required: [meta, scenario, edges, interventions, note]
      properties:
        meta: { $ref: "#/components/schemas/Meta" }
        scenario: { $ref: "#/components/schemas/Entity" }
        edges:
          type: array
          items: { $ref: "#/components/schemas/Entity" }
        interventions:
          type: array
          items: { $ref: "#/components/schemas/Entity" }
        note: { type: string }
    ForecastsResponse:
      type: object
      required: [meta, status, producer, forecasts, aggregations, estimates, note]
      properties:
        meta: { $ref: "#/components/schemas/Meta" }
        status: { type: string, const: external_aggregate }
        producer: { type: string }
        forecasts:
          type: array
          description: Forecast records with original question wording, outcome_set, horizon and conditions.
          items: { $ref: "#/components/schemas/Entity" }
        aggregations:
          type: array
          description: One row per compatibility group × method; each carries outcome_set, horizon and conditioning.
          items:
            type: object
            additionalProperties: true
        estimates:
          type: array
          items: { $ref: "#/components/schemas/Estimate" }
        note: { type: string }
    CapabilitiesResponse:
      type: object
      required: [meta, benchmarks, results]
      properties:
        meta: { $ref: "#/components/schemas/Meta" }
        benchmarks:
          type: array
          items: { $ref: "#/components/schemas/Entity" }
        results:
          type: array
          items: { $ref: "#/components/schemas/Entity" }
...
    IncidentsResponse:
      type: object
      required: [meta, incidents]
      properties:
        meta: { $ref: "#/components/schemas/Meta" }
        incidents:
          type: array
          items: { $ref: "#/components/schemas/Entity" }
    SafeguardsResponse:
      type: object
      required: [meta, interventions]
      properties:
        meta: { $ref: "#/components/schemas/Meta" }
        interventions:
          type: array
          items: { $ref: "#/components/schemas/Entity" }
    SourcesResponse:
      type: object
      required: [meta, total, limit, offset, sources]
      properties:
        meta: { $ref: "#/components/schemas/Meta" }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
        sources:
          type: array
          items: { $ref: "#/components/schemas/Entity" }
    SourceResponse:
      type: object
      required: [meta, source, claims]
      properties:
        meta: { $ref: "#/components/schemas/Meta" }
        source: { $ref: "#/components/schemas/Entity" }
        claims:
          type: array
          items: { $ref: "#/components/schemas/Entity" }
...
    ReleaseResponse:
      type: object
      required: [meta, is_current, release]
      properties:
        meta: { $ref: "#/components/schemas/Meta" }
        is_current: { type: boolean }
        release:
          type: object
          required: [manifest, estimates, indexes, aggregations, sensitivity, delta, drivers_explained, approvals]
          properties:
            manifest: { $ref: "#/components/schemas/ReleaseManifest" }
            estimates:
              type: array
              items: { $ref: "#/components/schemas/Estimate" }
            indexes:
              type: array
              items: { $ref: "#/components/schemas/IndexValue" }
            aggregations:
              type: array
              items: { type: object, additionalProperties: true }
            sensitivity:
              type: array
              items: { type: object, additionalProperties: true }
            delta: { type: object, additionalProperties: true }
            drivers_explained: { type: object, additionalProperties: true }
            approvals:
              type: array
              items: { type: object, additionalProperties: true }
...
    ActionsResponse:
      type: object
      required: [meta, actions]
      properties:
        meta: { $ref: "#/components/schemas/Meta" }
        actions:
          type: array
          items: { $ref: "#/components/schemas/Entity" }
...